pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "mlflow2prov"
version = "0.1.0"
description = "Extract provenance graphs compliant with W3C PROV from ML experiment projects that use Git repositories and MLflow tracking."
license = "Apache-2.0"
authors = ["Marius Schlegel <marius.schlegel@tu-ilmenau.de>"]
readme = "README.md"
repository = "https://github.com/mariusschlegel/mlflow2prov"
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
packages = [{ include = "mlflow2prov" }]

[tool.poetry.scripts]
mlflow2prov = "mlflow2prov.entrypoints.cli:cli"

[tool.poetry.dependencies]
python = "~3.10"
click = "^8.1.3"
gitpython = "^3.1.31"
jsonschema = "^4.17.3"
mlflow = "==2.3.0"
numba = "^0.56.4"
prov = { extras = ["dot"], version = "^2.0.0" }
ruamel-yaml = "^0.17.21"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
genbadge = { version = "^1.1.0", extras = ["coverage"] }
isort = "^5.12.0"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
minversion = "7.2"
addopts = "-rA --cov=mlflow2prov --cov-report term-missing --cov-report xml:assets/coverage.xml"
testpaths = ["tests"]
log_level = "WARN"
# log_cli = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise AssertionError",
    "raise NotImplementedError",
    "@abc.abstractmethod",
    "if __name__ == .__main__.:",
]
